+++ src/main/java/code/services/rooms/DeviceServiceImpl.java
@@ -1,7 +1,9 @@
 package code.services.rooms;
 
-import code.model.dto.rooms.DeviceRequestDTO;
-import code.model.dto.rooms.DeviceResponseDTO;
+import code.exception.BadRequestException;
+import code.exception.ResourceNotFoundException;
+import code.model.dto.rooms.req.DeviceRequestDTO;
+import code.model.dto.rooms.res.DeviceResponseDTO;
 import code.model.entity.rooms.DevicesEntity;
 import code.repository.rooms.DevicesRepository;
 import code.util.RandomId;
@@ -26,6 +28,9 @@ public class DeviceServiceImpl implements  DeviceService {
 
     @Override
     public boolean insertDevice(DeviceRequestDTO deviceRequestDTO) {
+        if (deviceRequestDTO == null){
+            throw new BadRequestException("DeviceRequestDTO is null");
+        }
         DevicesEntity devicesEntity = new DevicesEntity();
         devicesEntity.setDeviceId(generateDeviceId());
         devicesEntity.setDeviceName(deviceRequestDTO.getDeviceName());
@@ -37,9 +42,12 @@ public class DeviceServiceImpl implements  DeviceService {
 
     @Override
     public boolean updateDevice(String deviceId, DeviceRequestDTO deviceRequestDTO) {
+        if (deviceId == null || deviceRequestDTO == null){
+            throw new BadRequestException("DeviceId or DeviceRequestDTO is null");
+        }
         DevicesEntity devicesEntity = devicesRepository.findById(deviceId).orElse(null);
         if(devicesEntity == null){
-            return  false;
+            throw new ResourceNotFoundException("Device not found");
         }
         devicesEntity.setDeviceName(deviceRequestDTO.getDeviceName());
         devicesEntity.setDeviceType(deviceRequestDTO.getDeviceType());
@@ -50,9 +58,12 @@ public class DeviceServiceImpl implements  DeviceService {
 
     @Override
     public boolean deleteDevice(String deviceId) {
+        if (deviceId == null) {
+            throw new BadRequestException("DeviceId is null");
+        }
         DevicesEntity devicesEntity = devicesRepository.findById(deviceId).orElse(null);
         if(devicesEntity == null){
-            return  false;
+            throw new ResourceNotFoundException("Device not found");
         }
         devicesRepository.delete(devicesEntity);
         return true;
@@ -61,6 +72,9 @@ public class DeviceServiceImpl implements  DeviceService {
     @Override
     public List<DeviceResponseDTO> getAllDevices() {
         List<DevicesEntity> devicesEntities = devicesRepository.findAll();
+        if (devicesEntities == null){
+            throw new ResourceNotFoundException("Device List not found");
+        }
         return devicesEntities.stream().map(device -> new DeviceResponseDTO(
                 device.getDeviceId(),
                 device.getDeviceName(),
@@ -70,11 +84,17 @@ public class DeviceServiceImpl implements  DeviceService {
 
     @Override
     public DevicesEntity getDeviceById(String deviceId) {
+        if (deviceId == null) {
+            throw new BadRequestException("DeviceId is null");
+        }
         return devicesRepository.findById(deviceId).orElse(null);
     }
 
     @Override
     public DeviceResponseDTO getDeviceByDeviceId(String deviceId) {
+        if (deviceId == null) {
+            throw new BadRequestException("DeviceId is null");
+        }
         DevicesEntity devicesEntity = devicesRepository.findById(deviceId).orElse(null);
         if(devicesEntity == null){
             return null;
